able to draw path lines on the map

+++ public/js/events.js
@@ -12,7 +12,7 @@ userNameInput = document.getElementById("username-input"),
 passwordInput = document.getElementById("password-input"),
 logonButton = document.getElementById("logon-button"),
 messageBar = document.getElementById("percent"),
-notificationLabel = document.getElementById("notification"),
+notificationContainer = document.getElementById("notification"),
 removeNotification = document.getElementById("remove-notification"),
 quickplayNotification = document.getElementById("quickplay-notification"),
 notificationSelector = document.getElementById("notification-selector");
@@ -73,7 +73,6 @@ logonButton.onclick = function(){
 document.getElementById("submit-shared-file").addEventListener('click',function(){
 
 	var userLabelValue = document.getElementById("user-input").value;
-	console.log("user-input: "+userLabelValue);
 	if(userLabelValue!=null && userLabelValue!=currentUser.name && selectedExploration!=null){
 		saveFileToSharedUser(userLabelValue);
 		var selectedExplName = selectedExploration.name;
@@ -81,10 +80,14 @@ document.getElementById("submit-shared-file").addEventListener('click',function(
 	}
 });
 //notifications
-notificationLabel.addEventListener('click',function(){
-	notificationLabel.display = "none";
+notificationContainer.addEventListener('click',function(){
 	stopRecording();
-	showListNotifications();
+	if(showListNotifications()){
+		removeNotification.style.visibility = "visible";
+		quickplayNotification.style.visibility = "visible";
+
+	}
+
 });
 
 //new account
@@ -96,6 +99,12 @@ newAccount.onclick = function(){
 
 //remove current choice exploration
 deleteExplButton.onclick = function(){
+	if(selectedExploration==null){
+		return;
+	}
+	if(currentUser.explorations.indexOf(selectedExploration)<0){
+		return;
+	}
 	deleteExploration(selectedExploration);
 	deselectExploration();
 };
@@ -104,13 +113,20 @@ removeNotification.addEventListener("click", function(){
 	var selected = currentUser.getSharedExploration()[notificationSelector.options[notificationSelector.selectedIndex].value];
 	selected.isNew = false;
 	setExplorationIsOld(selected);
-	notificationSelector.style.display = "none";
-	removeNotification.style.display = "none";
-	quickplayNotification.style.display = "none";
+	notificationSelector.style.visibility = "hidden";
+	removeNotification.style.visibility = "hidden";
+	quickplayNotification.style.visibility = "hidden";
 	updateNotifications();
+	deselectExploration();
 });
 quickplayNotification.addEventListener("click", function(){
-	playExploration(currentUser.getCurrentExploration());
+	selected = currentUser.getSharedExploration()[notificationSelector.options[notificationSelector.selectedIndex].value];
+	var quickPlayExploration = new Exploration();
+	quickPlayExploration = selected;
+	//quickPlayExploration.events = selected.events;
+	//console.log(quickPlayExploration);
+	startPlayback(quickPlayExploration);
+	//selected.isNew = ture;
 });
 
 resetExplorations();
